Removed unuseful redirect to password reset.
The route removed redirects to the front area but not useful when we are in the admin area.
Changed to redirect to the "passwordreset" in the admin area. I made the "/request" in the URL optional so he can access the same page by "passwordreset" or "/passwordreset/request" both are the same.